requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-prometheus==2.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
prometheus_client==0.20.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
wheel==0.43.0
